New plugin releases not reported to updates.jenkins.io
The web page at https://updates.jenkins.io/latest/ has not been
updated in the last 31 hours. It is missing the most recent release
of the tuleap-api-plugin that was made about 23 hours ago. The link
https://updates.jenkins.io/latest/tuleap-api.hpi should point to
tuleap-api-plugin 2.3.0 but instead points to tuleap-api-plugin 2.2.2.

The update center job trying to update the page reports a failure when
downloading that plugin from artifactory. However, when I download
interactively using the same URL, it succeeds.