Run container as current user to avoid git issues
https://github.com/klakegg/docker-hugo/issues/82 notes that root is the
default user of the container, but that causes issues for our environment
because it blocks git operations.

Switch to run as the current user so that git operations are not
blocked.

Running as the current user means that the Hugo build lock cannot be
written.  Disable it with a command line switch.

Not a perfect solution, but simple enough to allow upgrade to a more
recent Hugo version.